corrected unlink(2) binding

+++ src/runtcl.c
@@ -3123,20 +3123,19 @@ static int objcmd_unlink ( ClientData cd, Tcl_Interp * const T,
   const int objc, Tcl_Obj * const * objv )
 {
   if ( 1 < objc ) {
-    int i ;
+    int i, j ;
+    const char * path = NULL ;
 
-    for ( i = 2 ; objc > i ; ++ i ) {
-      const char * path = Tcl_GetStringFromObj ( objv [ i ], NULL ) ;
+    for ( i = 1 ; objc > i ; ++ i ) {
+      j = -1 ;
+      path = Tcl_GetStringFromObj ( objv [ i ], & j ) ;
 
-      if ( path && * path ) {
+      if ( ( 0 < j ) && path && * path ) {
         if ( unlink ( path ) ) {
-          Tcl_SetErrno ( errno ) ;
-          Tcl_AppendResult ( T, "unlink \"", path, "\" failed: ",
-            Tcl_PosixError ( T ), (char *) NULL ) ;
-          return TCL_ERROR ;
+          return psx_err ( T, errno, "unlink" ) ;
         }
       } else {
-        Tcl_AddErrorInfo ( T, "empty path" ) ;
+        Tcl_AddErrorInfo ( T, "illegal file name" ) ;
         return TCL_ERROR ;
       }
     }
